ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
ultralytics==8.1.0
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.3
python-multipart==0.0.6
